+++ contracts/interface/IERC20Details.sol
@@ -4,5 +4,5 @@ pragma solidity 0.7.0;
 interface IERC20Details {
     function symbol() external view returns (string memory);
     
-    function decimals() external view returns (uint);
+    function decimals() external view returns (uint256);
 }
