+++ contracts/interface/IRepaymentCalculator.sol
@@ -2,5 +2,5 @@
 pragma solidity 0.7.0;
 
 interface IRepaymentCalculator {
-    function getNextPayment(address) external view returns (uint, uint, uint);
+    function getNextPayment(address) external view returns (uint256, uint256, uint256);
 }
